refactor(debugger): avoid JS debugger keyword conflict
