+++ packages/debugger/src/index.ts
@@ -1,3 +1,4 @@
-export { CKBDebugger } from "./debugger";
+export { CKBDebugger } from "./executor";
 export { createTestContext } from "./context";
 export { CKBDebuggerDownloader } from "./download";
+export * from "./types";
